backend/requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
Werkzeug==3.1.3
Flask==3.1.1
Flask-Cors==4.0.1
torch>=2.3.0
torchvision>=0.18.0
torchaudio>=2.3.0
transformers>=4.41.0
tokenizers>=0.19.0
safetensors>=0.4.2
accelerate>=0.31.0
flask-cors>=4.01
numpy>=1.24
tqdm>=4.65
kiwipiepy==0.21.0
kiwipiepy_model==0.21.0
colorama>=0.4.6
opencv-python-headless>=4.10.0.84
easyocr>=1.7.1
Pillow>=10.4.0
numpy>=1.26.0
imutils>=0.5.4
google-generativeai
python-dotenv